pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "topogps"
version = "0.1.0"
description = "TopoGPS: Topologically Grounded Internal GPS for Semantic Memory Retrieval (prototype)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Asser Aldardiri"}]
license = {text = "MIT"}
keywords = ["cognitive-architecture", "retrieval", "semantic-memory", "faiss", "networkx", "umap"]

dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "scikit-learn>=1.3",
  "networkx>=3.2",
  "sentence-transformers>=2.6",
  "faiss-cpu>=1.7.4",
  "umap-learn>=0.5.5",
  "plotly>=5.18",
  "pydantic>=2.6",
  "typer>=0.12",
  "rich>=13.7",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]

[project.scripts]
topogps = "topogps.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
